agent/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-chroma
python-dotenv
chromadb
langchain-text-splitters
beautifulsoup4
requests